Identified application using mimeType. Needs more testing
git-svn-id: file:///home/svn/shoola-omero/trunk@6997 6f2cb1de-eb0d-0410-b157-e593188b5901

+++ SRC/org/openmicroscopy/shoola/env/init/ContainerConfigInit.java
@@ -92,8 +92,10 @@ public final class ContainerConfigInit
     		File files;
             if (home.isDirectory()) files = new File(home, tmp);
             else files = new File(container.getHomeDir(), tmp);
-            files.mkdir();
-            files.deleteOnExit();
+            if (!files.exists()) {
+            	 files.mkdir();
+                 files.deleteOnExit();
+            }
 		} catch (ConfigException ce) {
 			throw new StartupException("Unable to load Container configuration",
 										ce);
